fix(ci): make the repo root a real gate, and fix what it found
The per-app workflows each scope clippy to one crate — imbib-rust.yml runs from
crates/imbib-core, imprint-rust.yml from crates/imprint-core. Nothing gated the
space between them, and two crates had been broken there for months with every
workflow green:

- impart-core did not compile with --features native at all. A crate-root
  `pub use types::Address` shadowed the `pub use ffi_types::*` glob below it, so
  the UDL scaffolding resolved the domain type where it needed the FFI record.
  Nothing outside the crate imported it; mbox.rs, which did rely on the root
  re-export, now names `types::Address` explicitly.
- impress-tags carried `uniffi::export` on parse_tag_query, whose return type is
  recursive through Box<TagQuery>. UniFFI has no lowering for Box<T>, so it
  never compiled either. Confirmed by adding derive(uniffi::Enum) and reading the
  error rather than assuming. Swift loses nothing — it parses tag filters itself
  in LocalFilterService.swift and has never called this.

New .github/workflows/workspace-rust.yml closes the gap: fmt --all, clippy
--workspace --all-targets --features native -D warnings, and a workspace test
run. The per-app workflows keep their app-specific steps. CLAUDE.md and AGENTS.md
state the two commands as definition-of-done for a Rust change.

Everything the new gate flagged is fixed, so it is green from the start:
impel-core had a dead assignment where the Active->Review step was never
asserted (now asserted rather than deleted), impress-git had \0 escapes followed
by digits, three assertions used .get(x).is_none(), and impress-tags' doc list
ran into the following paragraph. Green: workspace clippy clean under
-D warnings, workspace tests exit 0, imbib and imprint gates still clean.

Also fixes imbib's /api/status, which reported libraryCount: 0 on a 7-library
store: it used `searchLibrary(query: "").count`, and an empty query matches
nothing — and that was counting papers under a "library" name anyway. Now counts
libraries, collections, and adds publicationCount from the store's own COUNT(*).
That last field is why the route matters to agents: with no total available over
HTTP, count-publications had to sum per-library counts and silently missed the
152 publications filed in no library. The Rust client reads the field when
present and keeps the old sum as a fallback, so a new server still works against
an older running app — verified live, still 6379 there.

+++ crates/impress-llm/src/provider.rs
@@ -397,7 +397,7 @@ async fn complete_async(request: &LLMRequest) -> Result<LLMResponse, LLMError> {
     let content = response.text().unwrap_or_default().to_string();
 
     // Extract usage info if available
-    let tokens_used = response.usage().map(|u| u.total_tokens as u32);
+    let tokens_used = response.usage().map(|u| u.total_tokens);
 
     Ok(LLMResponse {
         content,
